fix(refactor): various refactorings - utc config key - cache package - lowercase @ errors - tried one command rule them all strategy. not suitable for GET commands - go-dms:v2
