+++ geocoding/service_client.go
@@ -8,8 +8,9 @@ import (
 	"net/http"
 	"strings"
 
-	dms "github.com/ddefrancesco/go-dms/dms"
-	cache "github.com/ddefrancesco/scoperunner_server/geocoding/cache"
+	dms "github.com/ddefrancesco/go-dms/v2/dms"
+
+	cache "github.com/ddefrancesco/scoperunner_server/cache"
 	"github.com/spf13/viper"
 )
