+++ go.mod
@@ -12,6 +12,7 @@ require (
 	github.com/Khan/genqlient v0.6.0 // indirect
 	github.com/adrg/xdg v0.4.0 // indirect
 	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
+	github.com/ddefrancesco/go-dms/v2 v2.0.0 // indirect
 	github.com/fsnotify/fsnotify v1.7.0 // indirect
 	github.com/hashicorp/hcl v1.0.0 // indirect
 	github.com/magiconair/properties v1.8.7 // indirect
@@ -40,7 +41,7 @@ require (
 require (
 	dagger.io/dagger v0.9.5
 	github.com/creack/goselect v0.1.2 // indirect
-	github.com/ddefrancesco/go-dms v1.0.2
+	github.com/ddefrancesco/go-dms/v2 v2.0.0
 	github.com/gocarina/gocsv v0.0.0-20231116093920-b87c2d0e983a
 	github.com/spf13/viper v1.18.2
 	github.com/stretchr/testify v1.9.0
