+++ scope-server-config.yaml
@@ -2,7 +2,7 @@ environments:
   fakescope: true
 serialport: "/dev/cu.usbserial-0001"
 timezone: "Europe/Rome"
-utc: "-01.0"
+utc-offset: "-01.0"
 geocoding:
   url: "https://geocoding.openapi.it/geocode"
   api_key:
